[REFACTORING] Split IMAPServerSequentialExecutionTest
